requirements.txt
fastapi
uvicorn
firebase-admin==6.4.0
google-adk==1.0.0
